# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chatterbox TTS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/presets": {
      "get": {
        "summary": "List Presets",
        "operationId": "list_presets_v1_presets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Preset",
        "operationId": "create_preset_v1_presets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/presets/{preset_id}": {
      "get": {
        "summary": "Get Preset",
        "operationId": "get_preset_v1_presets__preset_id__get",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Preset",
        "operationId": "update_preset_v1_presets__preset_id__put",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Preset",
        "operationId": "delete_preset_v1_presets__preset_id__delete",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/prep/adjust": {
      "post": {
        "summary": "Prep Adjust",
        "operationId": "prep_adjust_v1_prep_adjust_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepAdjustRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepAdjustResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/prep/interpret": {
      "post": {
        "summary": "Prep Interpret",
        "operationId": "prep_interpret_v1_prep_interpret_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepInterpretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepInterpretResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/audio/edit": {
      "post": {
        "summary": "Edit Audio",
        "operationId": "edit_audio_v1_audio_edit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioEditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AudioEditRequest": {
        "properties": {
          "input_wav_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Wav Path"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "trim_enabled": {
            "type": "boolean",
            "title": "Trim Enabled",
            "default": false
          },
          "normalize_enabled": {
            "type": "boolean",
            "title": "Normalize Enabled",
            "default": false
          },
          "target_dbfs": {
            "type": "number",
            "title": "Target Dbfs",
            "default": -1.0
          }
        },
        "type": "object",
        "title": "AudioEditRequest"
      },
      "AudioEditResponse": {
        "properties": {
          "edited_wav_path": {
            "type": "string",
            "title": "Edited Wav Path"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "edited_wav_path",
          "metrics"
        ],
        "title": "AudioEditResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrepAdjustRequest": {
        "properties": {
          "text_raw": {
            "type": "string",
            "title": "Text Raw",
            "default": ""
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          }
        },
        "type": "object",
        "title": "PrepAdjustRequest"
      },
      "PrepAdjustResponse": {
        "properties": {
          "text_adjusted": {
            "type": "string",
            "title": "Text Adjusted"
          }
        },
        "type": "object",
        "required": [
          "text_adjusted"
        ],
        "title": "PrepAdjustResponse"
      },
      "PrepInterpretRequest": {
        "properties": {
          "text_adjusted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Adjusted"
          },
          "text_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Raw"
          },
          "glossary_enabled": {
            "type": "boolean",
            "title": "Glossary Enabled",
            "default": false
          },
          "glossary_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glossary Profile"
          },
          "glossary_options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Glossary Options"
          }
        },
        "type": "object",
        "title": "PrepInterpretRequest"
      },
      "PrepInterpretResponse": {
        "properties": {
          "text_interpreted": {
            "type": "string",
            "title": "Text Interpreted"
          },
          "applied_rules_summary": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Rules Summary"
          }
        },
        "type": "object",
        "required": [
          "text_interpreted"
        ],
        "title": "PrepInterpretResponse"
      },
      "PresetCreateRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PresetCreateRequest"
      },
      "PresetListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PresetListItem"
      },
      "PresetListResponse": {
        "properties": {
          "presets": {
            "items": {
              "$ref": "#/components/schemas/PresetListItem"
            },
            "type": "array",
            "title": "Presets"
          }
        },
        "type": "object",
        "required": [
          "presets"
        ],
        "title": "PresetListResponse"
      },
      "PresetMutationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "PresetMutationResponse"
      },
      "PresetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "state": {
            "$ref": "#/components/schemas/UIState"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state"
        ],
        "title": "PresetResponse"
      },
      "PresetUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "PresetUpdateRequest"
      },
...
      "UIState": {
        "properties": {
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "preparation": {
            "$ref": "#/components/schemas/UIStatePreparation"
          },
          "direction": {
            "$ref": "#/components/schemas/UIStateDirection"
          },
          "engine": {
            "$ref": "#/components/schemas/UIStateEngine"
          },
          "post": {
            "$ref": "#/components/schemas/UIStatePost"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "UIState"
      },
      "UIStateChunkPreview": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "est_duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Duration S"
          },
          "word_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          }
        },
        "type": "object",
        "required": [
          "index",
          "text"
        ],
        "title": "UIStateChunkPreview"
      },
      "UIStateChunkRange": {
        "properties": {
          "start": {
            "type": "integer",
            "title": "Start"
          },
          "end": {
            "type": "integer",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "UIStateChunkRange"
      },
      "UIStateDirection": {
        "properties": {
          "snapshot_text": {
            "type": "string",
            "title": "Snapshot Text",
            "default": ""
          },
          "chunk_markers": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Chunk Markers"
          },
          "chunk_ranges": {
            "items": {
              "$ref": "#/components/schemas/UIStateChunkRange"
            },
            "type": "array",
            "title": "Chunk Ranges"
          },
          "chunks_preview": {
            "items": {
              "$ref": "#/components/schemas/UIStateChunkPreview"
            },
            "type": "array",
            "title": "Chunks Preview"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "UIStateDirection"
      },
      "UIStateEngine": {
        "properties": {
          "engine_id": {
            "type": "string",
            "title": "Engine Id",
            "default": ""
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "chatterbox_gap_ms": {
            "type": "integer",
            "title": "Chatterbox Gap Ms",
            "default": 0
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "UIStateEngine"
      },
      "UIStatePost": {
        "properties": {
          "edit_enabled": {
            "type": "boolean",
            "title": "Edit Enabled",
            "default": false
          },
          "trim_enabled": {
            "type": "boolean",
            "title": "Trim Enabled",
            "default": false
          },
          "normalize_enabled": {
            "type": "boolean",
            "title": "Normalize Enabled",
            "default": false
          },
          "target_dbfs": {
            "type": "number",
            "title": "Target Dbfs",
            "default": -1.0
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "UIStatePost"
      },
      "UIStatePreparation": {
        "properties": {
          "text_raw": {
            "type": "string",
            "title": "Text Raw",
            "default": ""
          },
          "text_adjusted": {
            "type": "string",
            "title": "Text Adjusted",
            "default": ""
          },
          "text_interpreted": {
            "type": "string",
            "title": "Text Interpreted",
            "default": ""
          },
          "glossary_enabled": {
            "type": "boolean",
            "title": "Glossary Enabled",
            "default": false
          },
          "glossary_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glossary Profile"
          },
          "glossary_options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Glossary Options"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "UIStatePreparation"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}